security last thing

+++ produit-service/src/main/java/org/example/produitservice/entitie/Product.java
@@ -15,9 +15,9 @@ import lombok.*;
 public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
     private String name;
     private String description;
     private double price;
-    private long quantity;
+    private int quantity;
 }
